[#21] Refactor: loginApi를 authApi로 변경, 중복된 코드 삭제
- login에서만 쓰이지 않으니 authApi로 이름 변경
- handleSignIn에서도 토큰 처리를 하고 postSignIn에서도 토큰 처리를 하는 중복코드가 있어서 handleSignIn에서 중복된 코드 삭제
